requirements.txt
pytest
selenium
allure-pytest
openpyxl
pytest-html
